데이터 입력 Dialog 추가. (계속 구현중)
- 빈 (empty) 데이터 에서 추가 할 때 Exception 발생하는 문제 해결.
- UkeData 의 timeStamp 값을 long 에서 int 로 변경함.
- 데이터입력 Dialog 에 초기데이터 set/ 입력데이터 get 함수 추가 구현

+++ src/myproject/UkeData.java
@@ -28,7 +28,7 @@ public class UkeData {
 	public Note[] notes;
 
 	class Note {			// 음표(or chord) 1개의 데이터
-		public long	timeStamp;
+		public int	timeStamp;
 		public String	chordName;
 		public String	technic;
 		public String	tab[];			// 
@@ -36,7 +36,7 @@ public class UkeData {
 		public String	lyric;
 		
 		public void	Note() {			// 초기화 하기 위한 생성자
-			timeStamp = 0L;
+			timeStamp = 0;
 			chordName = "";
 			technic = "";
 			tab = new String[0];
